Added preference to toggle header sort on load
