+++ ginga/examples/configs/plugin_Header.cfg
@@ -1,0 +1,8 @@
+#
+# Header plugin preferences file
+#
+# Place this in file under ~/.ginga with the name "plugin_Header.cfg"
+
+# Sort header
+sortable = True
+#sortable = False
